+++ .eslintrc
@@ -1,5 +1,6 @@
 env:
   browser: true
+  jquery: true
   node: true
 
 globals:
@@ -7,5 +8,8 @@ globals:
 
 rules:
   global-strict: 0
-  no-use-before-define: [2, 'nofunc']
+  no-process-exit: 0
+  no-shadow: 0
+  no-unused-vars: [1, {args: none}]
+  no-use-before-define: [2, nofunc]
   quotes: [1, single]
