+++ client/static/assets/views/_partials/footer-nav.html
@@ -3,7 +3,7 @@
     <hr />
     <nav>
       <dl class="sub-nav right" role="menu">
-        <dt><strong>&copy; {{ YEAR }} <a href="https://mozilla.com/">Mozilla</a></strong></dt>
+        <dt><strong>&copy; {{ YEAR }} <a href="https://github.com/loads/loads-web">Mozilla</a></strong></dt>
 
         <!-- <dd><a ng-href="{{ ROUTES.HOME }}">Home</a></dd> -->
         <!-- <dd><a ng-href="{{ ROUTES.TESTS }}">Tests</a></dd> -->
