+++ package.json
@@ -1,26 +1,27 @@
 {
   "name": "loads-web",
-  "description": "Testing Hapi w/ Angular routing",
+  "description": "Web site for the loads v2 tool",
   "version": "1.0.0",
+
   "author": "Mozilla (https://mozilla.org/)",
   "bugs": "https://github.com/loads/loads-web/issues",
   "dependencies": {
+    "awstypes": "1.0.0",
     "bower": "1.3.12",
     "convict": "0.6.1",
-    "hapi": "8.1.0",
+    "hapi": "8.2.0",
     "hapi-auth-cookie": "2.0.0",
-    "joi": "5.1.0",
-    "mockloadsresults": "pdehaan/MockLoadsResults",
-    "mysql": "2.5.4",
+    "joi": "6.0.0",
+    "mysql": "2.5.5",
     "nunjucks-hapi": "1.1.0",
     "remarkable": "1.6.0",
     "require-hapiroutes": "0.1.7",
-    "sequelize": "2.0.0-rc8",
+    "sequelize": "2.0.3",
     "stylus": "0.50.0"
   },
   "devDependencies": {
     "awsbox": "0.7.0",
-    "eslint": "0.15.0",
+    "eslint": "0.15.1",
     "nsp": "1.0.0"
   },
   "homepage": "https://github.com/loads/loads-web",
