+++ server/controllers/assets.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+  directory: {
+    path: 'client/static/assets',
+    listing: false,
+    index: false,
+    lookupCompressed: true
+  }
+};
