+++ server/controllers/bower_components.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+  directory: {
+    path: 'client/bower_components',
+    index: false,
+    listing: false,
+    lookupCompressed: true
+  }
+};
