+++ server/controllers/logout.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function (request, reply) {
+  request.auth.session.clear();
+  reply.redirect('/');
+};
