+++ server/routes/broker.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  method: 'GET',
+  path: '/api/{proxy*}',
+  handler: require('../controllers/proxy')
+};
