+++ server/routes/mockproject.json
@@ -1,0 +1,42 @@
+{
+  "plans": [
+    {
+      "steps": [
+        {
+          "step_name": "Plan 1 Name",
+          "instance_count": "1",
+          "run_max_time": "1",
+          "run_delay": "1",
+          "step_url": "",
+          "environment_data": "1 Environment Data",
+          "dns_mapping": "DNS 1",
+          "port_mapping": "Port 1",
+          "volume_mapping": "Volume 1",
+          "docker_series": "Docker 1"
+        }
+      ],
+      "plan_title": "Plan 1 Title",
+      "plan_description": "Plan 1 Description"
+    },
+    {
+      "steps": [
+        {
+          "step_name": "Plan 2",
+          "instance_count": "22",
+          "run_max_time": "22",
+          "run_delay": "22",
+          "step_url": "",
+          "environment_data": "Data 2",
+          "dns_mapping": "DNS 2",
+          "port_mapping": "Port 2",
+          "volume_mapping": "Volumne 2",
+          "docker_series": "Docer 2"
+        }
+      ],
+      "plan_title": "Plan 2",
+      "plan_description": "Description 2"
+    }
+  ],
+  "project_title": "P Title",
+  "project_id": "22"
+}
