+++ server/routes/project.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  method: 'POST',
+  path: '/mock/project',
+  handler: require('../controllers/project')
+};
